backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
firebase-admin>=6.2.0
python-dotenv>=1.0.0
groq>=0.9.0
websockets>=12.0
tree-sitter>=0.23.2
tree-sitter-python>=0.23.2
requests>=2.31.0
